Removed tube name from pool, updated readme
Originally when a pool member was acquired I wanted a tube to be
specified.  Pool members do not "reset" the used tube.  This feature
would prevent recycled connections from inappropriately using the tube
of the last caller.  After thinking through it, I felt it is better to
leave this to the user to manage, we are all grown ups.  If it is a
desired feature, it should occur on release, when we know the connection
is healthy.

I also updated the Readme.md to include an example of the pool usage.